// src/types/interfaces.ts
export interface Game {
    ctr: string;
    completed: number;
    game_status:string;
    ht_abbrev: string;
    at_abbrev: string;
    away_team: string;
    home_team: string;
    away_teamlogo: string;
    home_teamlogo: string;
    at_score: number;
    ht_score: number;
    gamedate: string;
    reftime: string;
    odds: number;
    probability_pct: number;
    period:number;
  } 